Split survey ResultQuestions component
